+++ spyne/server/_base.py
@@ -108,8 +108,8 @@ class ServerBase(object):
             return
 
         if ctx.out_document is None:
-            self.app.out_protocol.serialize(ctx,
-                                        message=self.app.out_protocol.RESPONSE)
+            ctx.out_protocol.serialize(ctx,
+                                        message=ctx.out_protocol.RESPONSE)
 
         if ctx.service_class != None:
             if ctx.out_error is None:
@@ -119,7 +119,7 @@ class ServerBase(object):
                 ctx.service_class.event_manager.fire_event(
                                             'method_exception_document', ctx)
 
-        self.app.out_protocol.create_out_string(ctx)
+        ctx.out_protocol.create_out_string(ctx)
 
         if ctx.service_class != None:
             if ctx.out_error is None: