+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog
 =========
 
+spyne-2.11.0
+------------
+ * Add altering output protocol by setting ctx.out_protocol
+
 spyne-2.10.3
 ------------
  * Add validation tests for HierDictDocument and fix seen issues.